v0.3.1022 - /models leaves client.ts
SCALE-SEAM ⑰. Nine health/QA/georef/federation methods move to
api/models.ts (withModels; model.ts remains /model). Four regions.
client.ts 3,412 → 3,353.
