+++ apps/web/src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "Massing",
-  "version": "0.3.1021",
+  "version": "0.3.1022",
   "identifier": "com.ibuilder.aecbim",
   "build": {
     "frontendDist": "../dist",
